Create clean_resample.py
The first step in preprocessing the data could be to clean the data by removing any missing or invalid values. Next, you could resample the data to a lower frequency (e.g., 0.1Hz) to reduce the amount of data and make it easier to work with. This could be done using methods such as downsampling or interpolation.